Refactor `setup.py` (#3393)
* Refactor setup.py

* Update doc with the new file path

* Rearrange spec_builders

* Update doc
